remove old plugin_initialization_service and rename new one
